requirements.txt
# FastAPI for the backend framework
fastapi==0.100.0

# ASGI server for running FastAPI
uvicorn==0.23.0

# MongoDB driver
pymongo==4.5.0

# Optional MongoDB asynchronous ODM
motor==3.3.0

# ChatGPT API integration
openai==0.27.0

# Redis client
redis==5.3.0

# Cloud storage interaction with AWS S3
boto3==1.28.0

# PDF/DOCX parsing (for extracting text from resumes)
PyPDF2==3.0.0
python-docx==0.8.11

# Data validation and serialization
pydantic==2.1.1

# PostgreSQL database support (if needed)
sqlalchemy==2.0.30
psycopg2-binary==2.9.6

# Password hashing and authentication
passlib[bcrypt]==1.7.4  # Passlib with bcrypt support

# Unit testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Linting and code quality
flake8==6.1.0

# Environment variable management
python-dotenv==1.0.0

# HTTP request handling (optional utility)
requests==2.31.0

# Docker interaction (optional for managing containers from Python)
docker==6.1.3

python-jose==3.3.0

